requirements.txt
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap4==1.1.1
django-crispy-forms==1.9.1
django-grappelli==2.14.2
django-heroku==0.3.1
django-import-export==2.1.0
django-phonenumber-field==4.0.0
django-tinymce4-lite==1.8.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
jsmin==2.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
phonenumbers==8.12.3
Pillow==7.1.2
psycopg2==2.8.5
pyenchant==3.0.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
